+++ zadachiIzychebnika/src/main/java/andrey/Zadacha44.java
@@ -1,0 +1,14 @@
+package andrey.glava2;
+
+/**
+ * Задача 4, стр. 93.
+ */
+public class Zadacha44 {
+    /**
+     * Метод main.
+     * @param args - аргумент метода main.
+     */
+    public static void main(String[] args) {
+        System.out.println("Один\nДва\nТри"); //Тут я использовал знак табуляци "\n".
+    }
+}
