--- zadachiIzychebnika/src/main/java/andrey/zadacha4.java
@@ -1,14 +1,0 @@
-package andrey.glava2;
-
-/**
- * Задача 4, стр. 93.
- */
-public class zadacha4 {
-    /**
-     * Метод main.
-     * @param args - аргумент метода main.
-     */
-    public static void main(String[] args) {
-        System.out.println("Один\nДва\nТри"); //Тут я использовал знак табуляци "\n".
-    }
-}
